% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DHC_PCT2}
\alias{DHC_PCT2}
\title{Glossary for table PCT2 in the Demographic and Housing Characteristics survey.}
\format{
a data frame with seven columns
\describe{
\item{Index}{<int> the row number of the variable}
\item{Level}{<int> the level of aggregation of the variable}
\item{One Person}{<lgl> single-person or multi-person household}
\item{Children}{<lgl> are children present in the household}
\item{Family}{<lgl> are the dwellers a family}
\item{Male Householder}{<lgl> is there a male head of household}
\item{Female Householder}{<lgl> is there a female head of household}
}
}
\source{
\url{https://api.census.gov/data/2020/dec/dhc/groups/PCT2.html}
}
\usage{
DHC_PCT2
}
\description{
Glossary for table PCT2 in the Demographic and Housing Characteristics survey.
}
\keyword{datasets}
