% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_variable_details.R
\name{fetch_variable_details}
\alias{fetch_variable_details}
\title{Download and parse the JSON representation of a decennial table's variables}
\usage{
fetch_variable_details(.vintage, .table_or_survey_code, .group = NULL)
}
\arguments{
\item{.vintage}{\verb{<int>} which decennial census the table comes from}

\item{.table_or_survey_code}{\verb{<chr>} the abbreviated name of the data product where the table appears, like "pl" or "dhc}

\item{.group}{\verb{<chr?>} the specific table to pull information about, like "PCT12" or "PL"}
}
\value{
\verb{<tbl>} a data frame with a "Variable" column and one or more columns of details.
}
\description{
If you don't ask for a specific table with the \code{.group} argument then this
function will download information about every variable in the data product.
}
