% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GLOSSARY_OF_AGE_BY_SEX}
\alias{GLOSSARY_OF_AGE_BY_SEX}
\title{Sex by age for selected age categories.}
\format{
\subsection{GLOSSARY_OF_AGE_BY_SEX}{

An object of class \code{tbl_df}/\code{tbl}/\code{data.frame}
with 49 rows and 4 columns
\describe{
\item{Index    }{\verb{<int>} the row of the table that corresponds to this sex-age combination.}
\item{Sex      }{\verb{<chr>} "All", "Male," or "Female".}
\item{Lower Age}{\verb{<int>} inclusive}
\item{Upper Age}{\verb{<int>} inclusive, with \code{999} as the value for "or over."}
}
}
}
\source{
\url{https://api.census.gov/data/2000/dec/sf1/groups/P012.html}

\url{https://api.census.gov/data/2010/dec/sf1/groups/P12.html}

\url{https://api.census.gov/data/2020/dec/dhc/groups/P12.html}
}
\usage{
GLOSSARY_OF_AGE_BY_SEX
}
\description{
There are 10 of these tables, including one for all people and nine for
people in different
\href{https://www.census.gov/library/visualizations/interactive/decennial-census-measurement-of-race-and-ethnicity-across-the-decades-1790-2020.html}{OMB97}
racial/ethnic categories.
The categories range from 1 to 5 or more years in width.
}
\keyword{datasets}
