% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GLOSSARY_OF_ETHNICITY_BY_RACE}
\alias{GLOSSARY_OF_ETHNICITY_BY_RACE}
\title{Combinations of racial and Hispanic or non-Hispanic ethnic identities.}
\format{
\subsection{GLOSSARY_OF_RACE_ETHNICITY_FOR_SINGLE_YEAR_AGE}{

A list of two data tables
\describe{
\item{Indices}{Correspondences between row number and race/ethnicity}
\item{Vintages}{The specific information needed to pull the data from each decennial census}
}
}
}
\source{
\url{https://api.census.gov/data/2000/dec/sf1/groups/P008.html}

\url{https://api.census.gov/data/2010/dec/sf1/groups/P5.html}

\url{https://api.census.gov/data/2020/dec/dhc/groups/P5.html}
}
\usage{
GLOSSARY_OF_ETHNICITY_BY_RACE
}
\description{
Each vintage provides a slightly different path to its block-level counts of
people by OMB race and ethnicity.
All three tables have the same structure, however.
That means there are two parts to this glossary:
}
\details{
\subsection{Indices}{

An object of class \code{tbl_df}/\code{tbl}/\code{data.frame}
with 13 rows and 3 columns
\describe{
\item{Index}{\verb{<int>} the row number of the variable}
\item{Ethnicity}{\verb{<chr>} either "Hispanic" or "Not Hispanic"}
\item{Race}{\verb{<chr>} One of the seven OMB racial categories}
}
}

\subsection{Vintages}{

An object of class \code{tbl_df}/\code{tbl}/\code{data.frame}
with 3 rows and 4 columns
\describe{
\item{Vintage}{\verb{<int>} a decennial census year}
\item{Product}{\verb{<chr>} a short code for building the API URL}
\item{Group}{\verb{<chr>} the code for the table}
\item{Separator}{\verb{<chr>} used for building variable names, either "" or "_"}
\item{Suffix}{\verb{<chr>} Later vintages have multiple pieces of information per index, distinguished by suffix}
}
}
}
\keyword{datasets}
