% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GLOSSARY_OF_RACE_ETHNICITY_FOR_SINGLE_YEAR_AGE}
\alias{GLOSSARY_OF_RACE_ETHNICITY_FOR_SINGLE_YEAR_AGE}
\title{Combinations of racial and Hispanic or non-Hispanic ethnic identities.}
\format{
\subsection{GLOSSARY_OF_RACE_ETHNICITY_FOR_SINGLE_YEAR_AGE}{

An object of class \code{tbl_df}/\code{tbl}/\code{data.frame}
16 rows and 4 columns
\describe{
\item{Race Code  }{\verb{<chr>} An empty string or a capital lettern from "A" to "O."}
\item{Race       }{\verb{<chr>} Brief labels for the racial identities of each table.}
\item{Hispanic   }{\verb{<lgl>} Whether the table counts people of Hispanic (\code{TRUE}), non-Hispanic (\code{FALSE}), or All ethnicities (\code{NA}).}
\item{Description}{\verb{<chr>} The full description, from the Census, of the race-ethnicity combination described by the table.}
}
}
}
\source{
\url{https://api.census.gov/data/2000/dec/sf1/groups/PCT012.html}

\url{https://api.census.gov/data/2010/dec/sf1/groups/PCT12.html}

\url{https://api.census.gov/data/2020/dec/dhc/groups/PCT12.html}
}
\usage{
GLOSSARY_OF_RACE_ETHNICITY_FOR_SINGLE_YEAR_AGE
}
\description{
There are fifteen tables that count people by exact age, sex.
Each different table documents the age-by-sex structure of a different
combination of racial and Hispanic or non-Hispanic ethnic identities.
}
\keyword{datasets}
