% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GLOSSARY_OF_SINGLE_YEAR_AGE_BY_SEX}
\alias{GLOSSARY_OF_SINGLE_YEAR_AGE_BY_SEX}
\title{A very detailed accounting of people's ages by sex}
\format{
\subsection{GLOSSARY_OF_SINGLE_YEAR_AGE_BY_SEX}{

An object of class \code{tbl_df}/\code{tbl}/\code{data.frame}
with 209 rows and 4 columns
\describe{
\item{Sex      }{\verb{<chr>} NA, "Male," or "Female".}
\item{Lower Age}{\verb{<int>} inclusive, because the extreme age groups are ranges.}
\item{Upper Age}{\verb{<int>} exclusive. because the extreme age groups are ranges.}
\item{Index    }{\verb{<int>} the row of the table that corresponds to this sex-age combination.}
}
}
}
\source{
\url{https://api.census.gov/data/2000/dec/sf1/groups/PCT012.html}

\url{https://api.census.gov/data/2010/dec/sf1/groups/PCT12.html}

\url{https://api.census.gov/data/2020/dec/dhc/groups/PCT12.html}
}
\usage{
GLOSSARY_OF_SINGLE_YEAR_AGE_BY_SEX
}
\description{
There are fifteen tables that count people by exact age, sex.
They all have the same format, with the same rows in each table corresponding
to the same ages and sexes.
Each different table documents the age-by-sex structure of a different race
or ethnic group
This particular collection has a detailed partitioning of racial identities
into Hispanic and non-Hispanic ethnic identities.
}
\keyword{datasets}
