% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GLOSSARY_OF_TENURE_BY_ETHNICITY_AND_RACE}
\alias{GLOSSARY_OF_TENURE_BY_ETHNICITY_AND_RACE}
\title{Number of households by owning/renting by Hispanic/not by OMB race}
\format{
description
An object of class \code{tbl_df}/\code{tbl}/\code{data.frame}
with 93 rows and 7 columns
\describe{
\item{Vintage}{\verb{<int>} which decennial census}
\item{Group}{\verb{<chr>} the table in the decennial census product}
\item{Index}{\verb{<int>} the row in the table}
\item{Variable}{\verb{<chr>} the full munged name of the variable}
\item{Tenure}{\verb{<chr>} whether the households are owners or renters}
\item{Ethnicity}{\verb{<chr>} whether the householders are Hispanic or not}
\item{Race}{\verb{<chr>} which OMB race the householders identify as}
}
}
\source{
\url{https://api.census.gov/data/2000/dec/sf3/variables.html}

\url{https://api.census.gov/data/2010/dec/sf1/groups/HCT1.html}

\url{https://api.census.gov/data/2020/dec/dhc/groups/HCT1.html}
}
\usage{
GLOSSARY_OF_TENURE_BY_ETHNICITY_AND_RACE
}
\description{
Number of households by owning/renting by Hispanic/not by OMB race
}
\keyword{datasets}
