% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hercdectables-package.R
\docType{package}
\name{hercdectables-package}
\alias{hercdectables}
\alias{hercdectables-package}
\title{hercdectables: Glossaries that Explain the Census API's Decennial Group Tables}
\description{
A data package that contains glossaries for US Census groups. A glossary is a mapping from variable names to a collection of explicit values that the variables encode. For example, a variable in the Census database might report a count of the number of people from one demographic group living in one type of housing and primarily speaking one language. The glossary for that group would include columns of demographic group, housing type, and primary language. These glossaries explain the Census API's Decennial group tables.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://higherx4racine.github.io/hercdectables/}
}

}
\author{
\strong{Maintainer}: Ben Taft \email{ben.taft@career2cradle.org}

Other contributors:
\itemize{
  \item Higher Expectations for Racine County [copyright holder]
}

}
\keyword{internal}
