% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoist_group_details.R
\name{hoist_group_details}
\alias{hoist_group_details}
\title{Unpack the ragged list of details about a group's variables}
\usage{
hoist_group_details(.glossary, .detail_col = "Details")
}
\arguments{
\item{.glossary}{\verb{<tbl>} A data frame (probably \code{\link{DHC_VARIABLES}})}

\item{.detail_col}{\verb{<chr?>} The name of the column with ragged lists of information. Defaults to "Details"}
}
\value{
\verb{<tbl>} a data frame with extra columns hoisted from \code{.detail_col}
}
\description{
The table that this function creates will have one column for each value in
the longest list among \code{.glossary}'s values of \code{.detail_col}.
They will be labeled with capital letters.
That means that the function will crash in the unlikely case where the
widest element of  \code{.glossary}'s \code{.detail_col} has more than 26 values.
}
