% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syabsbre.R
\name{syabsbre}
\alias{syabsbre}
\title{Fetch population data broken down by single-year age, sex, and race/ethnicity}
\usage{
syabsbre(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[hercacstables:fetch_data]{hercacstables::fetch_data}}
  \describe{
    \item{\code{for_geo}}{<chr> the geographical level the data will describe, e.g. \code{"tract"}}
    \item{\code{for_items}}{<chr[]> one or more instances of \code{for_geo} desired, e.g. \code{"*"} or \code{"000200"}, passed on to \code{\link[hercacstables:build_for_geographies]{build_for_geographies()}}}
    \item{\code{use_key}}{<lgl?> optional, should the query include a Census API key from the system environment. Defaults to \code{TRUE}}
  }}
}
\value{
a table with 8 columns
\describe{
\item{Vintage}{\verb{<int>} the decennial census that is the source of this row}
\item{Race}{\verb{<chr>} a short description of the race of people counted by this row}
\item{Hispanic}{\verb{<lgl>} whether or not this row's people are Hispanic (\code{NA} means "both.")}
\item{Sex}{\verb{<chr>} female or male}
\item{Lower Age}{\verb{<int>} the lowest age among people counted by this row}
\item{Upper Age}{\verb{<int>} the greatest age among people counted by this row}
\item{GEO_ID}{\verb{<chr>} the alphanumeric designation of the loction described by this row}
\item{Population}{\verb{<chr>} the count of people with all of the characteristics in the other columns.}
}
}
\description{
Fetch population data broken down by single-year age, sex, and race/ethnicity
}
