% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DHC_VARIABLES}
\alias{DHC_VARIABLES}
\title{Variables from Demographic and Housing Characteristics survey.}
\format{
a data frame with five columns
\describe{
\item{Concept}{<chr> a verbal description of the group's contents}
\item{Group}{<chr> the code that designates the variable's group}
\item{Index}{<int> the row number of the variable}
\item{Variable}{<chr> the complete variable name}
\item{Details}{<list> the details for this variable}
}
}
\source{
\url{https://api.census.gov/data/2020/dec/dhc/variables.html}
}
\usage{
DHC_VARIABLES
}
\description{
Variables from Demographic and Housing Characteristics survey.
}
\keyword{datasets}
